Fix remaining CI failures: STILL_ACTIVE cast, clipboard test, env var race
- Cast STILL_ACTIVE to u32 in 3 Windows lock files (i32 vs u32 mismatch)
- Mark test_clip_by_id_succeeds #[ignore] (no clipboard in CI)
- Add LibraryManager::with_config_dir for tests
- Remove unsafe set_var calls from validate tests (race on macOS CI)

+++ src/auto_sync/lock.rs
@@ -141,7 +141,7 @@ pub fn process_alive(pid: u32) -> bool {
         if ok == 0 {
             return true;
         }
-        exit_code == STILL_ACTIVE
+        exit_code == STILL_ACTIVE as u32
     }
 }
